requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
joblib==1.3.1
xgboost==1.7.6
prophet==1.1.4
statsmodels==0.14.0
flask==2.3.2
gunicorn==21.2.0
pytest==7.4.0
black==23.7.0
flake8==6.0.0 